# openapi.yaml
openapi: 3.0.3
info:
  title: Bambuzle API
  description: |
    REST API for the Bambuzle 3D printer monitoring dashboard.

    Bambuzle connects to BambuLab printers via MQTT, collects telemetry (temperatures, layer progress, fan speeds),
    tracks print jobs, detects anomalies, and provides real-time state via WebSocket.

    ## Authentication
    Two separate layers:
    1. **Dashboard admin token** (who may use this API). Every `/api/*` route and the WebSocket require it, except
       `/api/session`, `/api/spec` and `/api/docs` (and the `/healthz` / `/readyz` probes, which live outside `/api`). Send `Authorization: Bearer <token>`, or exchange the token for an
       HttpOnly session cookie via `POST /api/session`. `BAMBUZLE_PUBLIC_READ=true` opens GETs and the WebSocket;
       `BAMBUZLE_AUTH=off` disables the check entirely.
    2. **BambuLab Cloud login** (the server's own connection to your printers), via `/api/auth/login`.

    ## Real-time Data
    For live updates, connect to the WebSocket endpoint at `ws://<host>:<port>/ws`.
    Messages are JSON with `{ type, data }` format. Types include `state`, `event`, and `auth`.
  version: 0.7.0
  contact:
    name: Bambuzle
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
  - sessionCookie: []
...
paths:
...
  /api/printers/{deviceId}/connection:
    get:
      tags: [Printers]
      summary: Connection settings + capabilities (access code is never returned — only hasAccessCode)
      parameters: [{ $ref: '#/components/parameters/DeviceId' }]
      responses: { '200': { description: Connection view }, '404': { description: Unknown printer } }
    put:
      tags: [Printers]
      summary: Update connection mode / LAN host / access code (omit accessCode to keep, '' to clear)
      parameters: [{ $ref: '#/components/parameters/DeviceId' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                mode: { type: string, enum: [auto, cloud, lan] }
                lanHost: { type: string, description: IPv4 or hostname }
                accessCode: { type: string, description: 8-character LAN access code (write-only) }
      responses: { '200': { description: Updated }, '400': { description: Invalid input }, '404': { description: Unknown printer } }
  /api/printers/{deviceId}/connection/test:
    post:
      tags: [Printers]
      summary: Test a LAN connection (saved settings unless overridden); reports the failing stage
      parameters: [{ $ref: '#/components/parameters/DeviceId' }]
      responses:
        '200':
          description: 'Result: { ok, stage: connected|auth|identity|tls|unreachable|error, message, developerMode }'

  # ─── Session (dashboard admin token) ───
...
  /api/audit:
    get:
      tags: [Audit]
      summary: Operator audit trail, newest first
      description: |
        BAM-41. Append-only record of security-relevant and state-changing operator actions: dashboard sign-in /
        sign-out / rate-limit hits, admin-guard denials on mutating requests (throttled to one row per source IP +
        action per minute; the next row carries `detail.suppressed`), BambuLab Cloud login / verify / logout,
        connection settings, LAN printer add / remove, connection tests, printer commands, alert rule and
        maintenance task changes, manual backups.

        Private even with `BAMBUZLE_PUBLIC_READ=true`. `detail` never contains secrets (access codes, passwords,
        verification codes, tokens, notifier config) — only field names and flags such as `accessCodeChanged`.
        Rows older than `audit.retentionDays` (default 365, env `BAMBUZLE_AUDIT_RETENTION_DAYS`, 0 = forever) are
        pruned by the daily cleanup.
      parameters:
        - $ref: '#/components/parameters/AuditFrom'
        - $ref: '#/components/parameters/AuditTo'
        - $ref: '#/components/parameters/AuditAction'
        - $ref: '#/components/parameters/AuditResult'
        - name: limit
          in: query
          schema:
            type: integer
            default: 500
            minimum: 1
            maximum: 2000
      responses:
        '200':
          description: Matching entries (at most `limit`)
          content:
            application/json:
              schema:
                type: object
                properties:
                  window:
                    type: object
                    properties:
                      from: { type: string, format: date-time }
                      to: { type: string, format: date-time }
                  filters:
                    type: object
                    properties:
                      action: { type: string, nullable: true }
                      result: { type: string, nullable: true }
                  limit: { type: integer }
                  truncated:
                    type: boolean
                    description: True when `limit` entries were returned (there may be more)
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEntry'
        '400':
          description: Bad from/to/action/result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/MaintUnauthorized'
  /api/audit/export:
    get:
      tags: [Audit]
      summary: Download the audit trail as CSV
      description: |
        Same filters as `GET /api/audit`, newest first, at most 50,000 rows (`X-Bambuzle-Truncated: true` when
        more matched). RFC 4180 with a UTF-8 BOM and CRLF line endings; text cells starting with `= + - @`, tab or
        CR are prefixed with `'` (formula-injection guard). Columns: id, ts, actor, source_ip, user_agent, action,
        target, result, detail (JSON), request_id.
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [csv]
            default: csv
        - $ref: '#/components/parameters/AuditFrom'
        - $ref: '#/components/parameters/AuditTo'
        - $ref: '#/components/parameters/AuditAction'
        - $ref: '#/components/parameters/AuditResult'
      responses:
        '200':
          description: CSV attachment
          content:
            text/csv:
              schema:
                type: string
        '400':
          description: Bad format or filters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/MaintUnauthorized'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Dashboard admin token (BAMBUZLE_ADMIN_TOKEN or <data dir>/admin-token)
    sessionCookie:
      type: apiKey
      in: cookie
      name: bambuzle_session
  parameters:
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
      description: Printer device ID
...
    AuditFrom:
      name: from
      in: query
      description: Window start (ISO 8601 date or date-time; offset-less values are UTC). Default 30 days before `to`.
      schema:
        type: string
        example: '2026-09-01'
    AuditTo:
      name: to
      in: query
      description: Window end (ISO 8601; a bare date means the end of that UTC day). Default now.
      schema:
        type: string
    AuditAction:
      name: action
      in: query
      description: Exact action (`printer.command`) or category prefix (`printer` matches `printer.*`)
      schema:
        type: string
        example: printer.command
    AuditResult:
      name: result
      in: query
      schema:
        type: string
        enum: [ok, denied, rejected, error]
  responses:
...
    MaintUnauthorized:
      description: Missing or invalid admin token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    AuditEntry:
      type: object
      properties:
        id: { type: integer }
        ts: { type: string, format: date-time, description: ISO 8601 UTC with milliseconds }
        actor:
          type: string
          description: Credential that authorised the request — `admin-token` (Bearer), `session` (cookie) or `anonymous`. Per-user identities arrive with BAM-16.
        source_ip: { type: string, nullable: true, description: Client IP (honours BAMBUZLE_TRUST_PROXY) }
        user_agent: { type: string, nullable: true, maxLength: 200 }
        action:
          type: string
          description: |
            session.login, session.logout, session.rate_limited, access.denied, access.cross_origin,
            cloud.login, cloud.verify, cloud.logout, printer.connection.update, printer.connection.test,
            printer.add, printer.remove, printer.command, alert.create, alert.update, alert.delete,
            maintenance.task.create, maintenance.task.update, maintenance.task.delete, maintenance.task.done,
            maintenance.templates, system.backup
        target: { type: string, nullable: true, example: 'printer:01S00EXAMPLE01' }
        result:
          type: string
          enum: [ok, denied, rejected, error]
          description: ok = done; denied = authentication/authorisation refused; rejected = validation, state gate or printer refusal; error = failure (incl. unconfirmed commands)
        detail:
          type: object
          nullable: true
          description: Small secret-free JSON, e.g. `{"accessCodeChanged":true}` or `{"command":"pause","outcome":"confirmed"}`
        request_id: { type: string, nullable: true, description: Also sent as the `X-Request-Id` response header }
    Error:
      type: object
      properties:
        error:
          type: string
...